[HIPPEROS] Interface with the HIPPEROS API
Added flags to compile newlib with dynamic reentrency.
This means that every thread now has its own reentrency
structure that is found at runtime using the facilities
provided by HIPPEROS.

Offloaded the implementation of the syscall wrappers
to the HIPPEROS API. The implementation in sys/hipperos
is now mostly empty.