[HIPPEROS] Dummy syscalls in crt0.o
Add a dummy implementation of all the system calls supported by Maestro
in `crto.c`.
This file will be compiled to `crt0.o` (and won't be part of `libc.a`)
and it will be used for compiler checks (particularly useful for the
`libstdc++` compilation).

This trick is similar to the one used by *rtems*.

Implementation of KERNEL-554.